refactor api key auth
